Removed public visibility of ChunkBatchCmdRes
